ShentelWebMap/ev-geodjango-jt_mapsite/map_site/requirements.txt
django-filter~=22.1.0
djangorestframework-gis~=1.0.0
djangorestframework~=3.14.0
django~=4.1.7
psycopg2-binary~=2.9.0